+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- automatic discovery in home assistant:
+  - entity `button.[hostname]_logind_lock_all_sessions`
+  - entity `button.[hostname]_logind_poweroff`
+
 ### Changed
 - automatic discovery in home assistant:
   - replace component-based (topic: